Major refactor
I deleted a large amount of useless code, and moved some things around.
The repo isn't totally clean, but it's a lot better. Most of the
functionality was broken by this diff, but I will update / fix / improve
in the next diffs. I didn't want to go any further without a commit.